feat(NoticeAdmin) : 검색 API 개발 test(NoticeService) : 공지게시판 테스트코드 3종 작성 fix(Notice, Question, Wish) : 클래스 전반적 수정
### 변경점
- 공지 검색 기능 추가

### 버그 리포트
- 개발 편의를 위해 추가해본 enum 타입이 정상작동하지 않는 이슈가 있었음
- 이슈 1. 콘솔창에서 해당 enum 클래스를 찾지 못한다고 로그 출력
- 이슈 2. bit(n) 타입에 integer를 넣을 수 없어서 빌드 불가

- enum 타입 관련 코드에 수정을 하고 gradle clean 후 build 하니 정상 작동함

### 테스트
- createNotice 성공/예외케이스 : 성공
- searchNotice 성공케이스 : 성공

+++ src/main/java/sesac/bookmanager/admin/notice/data/NoticeCreateRequest.java
@@ -11,5 +11,5 @@ public class NoticeCreateRequest {
     private String title;
     private String content;
     private NoticeType type;
-    private Long adminId;
+    private Integer adminId;
 }
